functions split into separate files, imported and called in main file. pytest updates to come
